requirements.txt
asgiref==3.2.3
Django==3.0.3
django-crispy-forms==1.9.0
django-mssql-backend==2.7.0
pyodbc==4.0.30
pytz==2019.3
sqlparse==0.3.0
